Resolve two more opensource export issues.
* Move Python 3-only tests to py3/test_errors.py.
* Ignore setup.py in copybara config.

PiperOrigin-RevId: 225086882
