+++ pytype/tests/test_errors.py
@@ -1,7 +1,5 @@
 """Tests for displaying errors."""
 
-import re
-
 from pytype import file_utils
 from pytype.tests import test_base
 
@@ -1092,85 +1090,6 @@ class InPlaceOperationsTest(test_base.TargetIndependentTest):
     self.assertErrorLogIs(errors, [
         (1, "unsupported-operands",
          r"\+\=.*List.*int.*__iadd__ on List.*Iterable")])
-
-  def _testOp(self, op, symbol):
-    errors = self.CheckWithErrors("""\
-      class A(object):
-        def __%s__(self, x: "A"):
-          return None
-      def f():
-        v = A()
-        v %s 3  # line 6
-    """ % (op, symbol))
-    self.assertErrorLogIs(errors, [
-        (6, "unsupported-operands",
-         r"%s.*A.*int.*__%s__ on A.*A" % (re.escape(symbol), op))])
-
-  def testISub(self):
-    self._testOp("isub", "-=")
-
-  def testIMul(self):
-    self._testOp("imul", "*=")
-
-  def testIDiv(self):
-    errors = self.CheckWithErrors("""\
-      class A(object):
-        def __idiv__(self, x: "A"):
-          return None
-        def __itruediv__(self, x: "A"):
-          return None
-      def f():
-        v = A()
-        v /= 3  # line 8
-    """)
-    self.assertErrorLogIs(errors, [
-        (8, "unsupported-operands", r"\/\=.*A.*int.*__i(true)?div__ on A.*A")])
-
-  def testIMod(self):
-    self._testOp("imod", "%=")
-
-  def testIPow(self):
-    self._testOp("ipow", "**=")
-
-  def testILShift(self):
-    self._testOp("ilshift", "<<=")
-
-  def testIRShift(self):
-    self._testOp("irshift", ">>=")
-
-  def testIAnd(self):
-    self._testOp("iand", "&=")
-
-  def testIXor(self):
-    self._testOp("ixor", "^=")
-
-  def testIOr(self):
-    self._testOp("ior", "|=")
-
-  def testIFloorDiv(self):
-    self._testOp("ifloordiv", "//=")
-
-
-class MatrixOperationsTest(test_base.TargetPython3FeatureTest):
-  """Test matrix operations."""
-
-  def testMatMul(self):
-    errors = self.CheckWithErrors("def f(): return 'foo' @ 3")
-    self.assertErrorLogIs(errors, [
-        (1, "unsupported-operands",
-         r"\@.*str.*int.*'__matmul__' on str.*'__rmatmul__' on int")])
-
-  def testIMatMul(self):
-    errors = self.CheckWithErrors("""\
-      class A(object):
-        def __imatmul__(self, x: "A"):
-          pass
-      def f():
-        v = A()
-        v @= 3  # line 6
-    """)
-    self.assertErrorLogIs(errors, [
-        (6, "unsupported-operands", r"\@.*A.*int.*__imatmul__ on A.*A")])
 
 
 class NoSymbolOperationsTest(test_base.TargetIndependentTest):
